feat: ui improvements

+++ src/pages/EditItemPage.tsx
@@ -10,6 +10,7 @@ export function EditItemPage() {
   const { id } = useParams();
   const [item, setItem] = useState<ItemData | null>(null);
   const [history, setHistory] = useState([]);
+  const [isConfirmingDelete, setIsConfirmingDelete] = useState(false);
 
   const fetchData = useCallback(async () => {
     try {
@@ -46,6 +47,12 @@ export function EditItemPage() {
   };
 
   const deleteItem = async () => {
+    if (!isConfirmingDelete) {
+      setIsConfirmingDelete(true);
+      setTimeout(() => setIsConfirmingDelete(false), 3000);
+      return;
+    }
+
     const res = await fetch(`/api/items/${id}`, {
       method: "DELETE",
     });
@@ -55,30 +62,47 @@ export function EditItemPage() {
       navigate("/");
     } else {
       alert("错误");
+      setIsConfirmingDelete(false);
     }
   }
 
   return <>
     <NavBar />
-    {item &&
-      <>
-        <ItemForm title={`编辑项目`} initialData={item} onSubmit={handleSubmit} />
-        <button onClick={deleteItem}>删除</button>
-        <h2>记录</h2>
-        {history.map((entry: any, index) => (
-          <div key={index}>
-            <p>
-              <span>{new Date(entry.timestamp).toLocaleString()}</span>
-            </p>
-            <ul>
-              {Object.keys(entry.new_values).map((column, idx) => (
-                <li key={idx}>{columnNames[column]}: {entry.old_values[column] || "(空白)"} ⇾ {entry.new_values[column] || "(空白)"}</li>
-              ))}
-            </ul>
+    <div className="page">
+      {item &&
+        <>
+          <h1>编辑项目</h1>
+          <div className="module">
+            <ItemForm initialData={item} onSubmit={handleSubmit} />
           </div>
-        ))}
-        <p>{new Date((item as any).created_at).toLocaleString()}: 创建</p>
-      </>
-    }
+          <div className="module">
+            <h2>删除</h2>
+            <p>删除此项目将不可恢复 (需点两次确认)</p>
+            <button
+              onClick={deleteItem}
+              className={isConfirmingDelete ? "button-delete-confirm" : ""}
+            >
+              {isConfirmingDelete ? "再次点击以确认删除" : "删除"}
+            </button>
+          </div>
+          <div className="module">
+            <h2>记录</h2>
+            {history.map((entry: any, index) => (
+              <div key={index} className="history-entry">
+                <p>
+                  <span>{new Date(entry.timestamp).toLocaleString()}</span>
+                </p>
+                <ul>
+                  {Object.keys(entry.new_values).map((column, idx) => (
+                    <li key={idx}>{columnNames[column]}: {entry.old_values[column] || "(空白)"} ⇾ {entry.new_values[column] || "(空白)"}</li>
+                  ))}
+                </ul>
+              </div>
+            ))}
+            <p>{new Date((item as any).created_at).toLocaleString()}: 创建</p>
+          </div>
+        </>
+      }
+    </div>
   </>;
 }
